nsh: datapath/userland support for network service headers
This patch adds support for Network Service Headers (nsh) over VXLAN
as mentioned in [1]. Here changes are made to datapath to add nsh
headers whenever a vxlan port with destination port as 6633 (which is
IANA allocated port for nsh over vxlan) is created.

[1] http://tools.ietf.org/html/draft-quinn-sfc-nsh-07

(added a small script to install ovs with nsh support. Be warned it
 will remove old ovs settings and db.)